+++ examples/testing-network-shim-app/cypress/support/index.js
@@ -1,8 +1,8 @@
-import { enableNetworkShim } from '@dhis2/cypress-commands'
+import { enableNetworkShim, isCaptureMode } from '@dhis2/cypress-commands'
 import { resetDb } from '../../src/json-server/resetDb'
 
 afterEach(() => {
-    if (Cypress.env('dhis2_api_stub_mode') === 'CAPTURE') {
+    if (isCaptureMode()) {
         resetDb()
     }
 })