+++ examples/testing-network-shim-app/src/NamesList.js
@@ -1,5 +1,5 @@
-import React from 'react'
 import propTypes from 'prop-types'
+import React from 'react'
 
 export const NamesList = ({
     names,