+++ packages/cli/src/commands/install/enableAutoLogin.js
@@ -1,0 +1,8 @@
+module.exports.enableAutoLogin = ({ options, state }) => {
+    if (!options.enableAutoLogin) return state
+
+    return {
+        ...state,
+        support: [...state.support, 'enableAutoLogin'],
+    }
+}