+++ packages/cli/src/commands/install/mergePluginSetupCalls.js
@@ -1,10 +1,11 @@
-const { runCodemod } = require('../../utils/runCodemod.js')
 const path = require('path')
+const { runCodemod } = require('../../utils/runCodemod.js')
 
 /**
  * Copied from: https://stackoverflow.com/questions/1960473/get-all-unique-values-in-a-javascript-array-remove-duplicates
  */
-const onlyUnique = (value, index, self) => self.indexOf(value) === index
+const onlyUnique = ({ name }, index, self) =>
+    self.findIndex(curValue => curValue.name === name) === index
 
 module.exports.mergePluginSetupCalls = (plugins, paths) => {
     const pluginInits = plugins.filter(onlyUnique)