+++ packages/cli/src/commands/install/setLoginBackendUrl.js
@@ -1,4 +1,4 @@
-const inquirer = require('inquirer')
+const inquirer = require('@dhis2/cli-helpers-engine').inquirer
 
 module.exports.setLoginBackendUrl = async ({ options, state }) => {
     if (!options.setLoginBackendUrl) return state
@@ -18,7 +18,7 @@ module.exports.setLoginBackendUrl = async ({ options, state }) => {
         ...state,
         cypressEnvJson: {
             ...state.cypressEnvJson,
-            dhis2_base_url: envAnswers.dhis2Url,
+            dhis2BaseUrl: envAnswers.dhis2Url,
         },
     }
 }