+++ packages/cypress-commands/src/setups/enableAutoLogin.js
@@ -1,4 +1,11 @@
+import { setDhis2BaseUrlToLocalStorage } from '../helper/dhis2BaseUrl.js'
+import { isStubMode } from '../helper/networkMode.js'
+
 export const enableAutoLogin = () => {
+    if (isStubMode()) {
+        return
+    }
+
     before(() => {
         cy.login()
     })
@@ -10,7 +17,6 @@ export const enableAutoLogin = () => {
         // This ensures the app platform knows which URL to use even if
         // REACT_APP_DHIS2_BASE_URL is undefined It also ensures that the value
         // from the cypress env is used instead of REACT_APP_DHIS2_BASE_URL
-        const loginUrl = Cypress.env('dhis2_base_url')
-        localStorage.setItem('DHIS2_BASE_URL', loginUrl)
+        setDhis2BaseUrlToLocalStorage()
     })
 }