+++ packages/cypress-plugins/src/index.js
@@ -1,5 +1,5 @@
-const cucumberPreprocessor = require('./plugins/cucumberPreprocessor')
 const chromeAllowXSiteCookies = require('./plugins/chromeAllowXSiteCookies')
+const cucumberPreprocessor = require('./plugins/cucumberPreprocessor')
 const networkShim = require('./plugins/networkShim/index.js')
 
 module.exports = {