+++ packages/cypress-plugins/src/plugins/networkShim/utils.js
@@ -1,27 +1,24 @@
 const path = require('path')
 
 module.exports.isCaptureMode = function (env) {
-    return env.dhis2_api_stub_mode === 'CAPTURE'
+    return env.networkMode === 'capture'
 }
 
 module.exports.isStubMode = function (env) {
-    return env.dhis2_api_stub_mode === 'STUB'
+    return env.networkMode === 'stub'
 }
 
-module.exports.isDisabledMode = function (env) {
-    return !env.dhis2_api_stub_mode || env.dhis2_api_stub_mode === 'DISABLED'
+module.exports.isLiveMode = function (env) {
+    return !env.networkMode || env.networkMode === 'live'
 }
 
 module.exports.getFixturesDir = function ({ fixturesFolder, env }) {
-    return path.join(
-        fixturesFolder,
-        'network',
-        env.dhis2_server_minor_version.toString()
-    )
+    return path.join(fixturesFolder, 'network', env.dhis2ApiVersion.toString())
 }
 
 module.exports.getDefaultStaticResources = function () {
     return [
+        'system/info',
         'systemSettings/applicationTitle',
         'me',
         'dhis-web-commons/menu/getModules.action',